backend/requirements.txt
fastapi
uvicorn[standard]
pyserial
numpy
pandas
scikit-learn
opencv-python
python-multipart
